refactor(sparse): extract merge helpers into sparse_index/merge.rs (#378 review)
Addresses the latest Devin review observation on PR #621. The earlier
extraction of `mutable_segment.rs` explicitly cited the 500 NLOC
threshold as the motivating constraint. This PR added the k-way
merge logic (four helpers: `merge_sorted_runs`, `k_way_merge`,
`find_min_head_doc_id`, `advance_matching_runs`) as free functions
inside `inverted_index.rs`, nudging the file back toward the limit
(382 NLOC, well under 500 — but trending up).

Move them into a dedicated `sparse_index/merge.rs` module, mirroring
the `mutable_segment.rs` pattern:

- `inverted_index.rs` : 382 → 332 NLOC (−50).
- `merge.rs`          : 52 NLOC (4 functions, all CC ≤ 3).
- Total NLOC unchanged at 384 across the two files.

Visibility: `merge_sorted_runs` is `pub(super)` so only
`inverted_index.rs` (and future sparse_index siblings) can call it.
The three internal helpers stay module-private.

No behavioural change — the functions move verbatim. 62/62
sparse_index + 7/7 recall_validation still PASS. Lizard reports
zero thresholds exceeded across both files.
